+++ Modules/Product/Database/Migrations/2021_08_23_072836_create_product_colors_table.php
@@ -19,9 +19,9 @@ class CreateProductColorsTable extends Migration
             $table->foreignId('product_id')->constrained('products')->onUpdate('cascade')->onDelete('cascade');
             $table->decimal('price_increase', 20, 3)->default(0);
             $table->tinyInteger('status')->default(0);
-            $table->tinyInteger('sold_number')->default(0);
-            $table->tinyInteger('frozen_number')->default(0);
-            $table->tinyInteger('marketable_number')->default(0);
+            $table->smallInteger('sold_number')->default(0);
+            $table->smallInteger('frozen_number')->default(0);
+            $table->smallInteger('marketable_number')->default(0);
             $table->timestamps();
             $table->softDeletes();
         });
